Add Default Configuration
Currently in the default config the "enabled" options are ignored as the
default is true and false won't override true using mergo. In order to
make defaults useful (i.e. work without a conf file), the solution is to
change "enabled" to disabled and make the default false.

This also means changing the "use_ssl" option to something like
"be_unsafe".