bugfix: golang panic, runtime error
golang runtime error -> 'index out of range [0] with length 0'

From raw_message.go:231. After checking if the 'data' slice length check
we need to be return if any errors occur. The previous behavior was to
try to index the slice. If the slice had len(data) == 0, this would
obvious fail.