Address Claim for CANUSB
CANUSB can now send ISO Address Claim PGNs and automatically renegotiate
when some other device wants to use it's address. It's likely that the
official J1939 and/or NMEA 2000 specification dictates who gets to keep
their address when there is a conflict. Since we don't know what those
rules are, whenever we see an address conflict, we try a new one.

Ideally, we'd also keep track of what addresses are in use and not
attempt to claim one of those. This code does not do that.

As a side effect of this update, we also implement a generic Send()
command for publishing data on the bus.